feat: 对齐数据格式

+++ data/view/cnf/def.go
@@ -12,21 +12,21 @@ var EImportsHead = map[string]string{
 
 // TypeMysqlDicMp Accurate matching type.精确匹配类型
 var TypeMysqlDicMp = map[string]string{
-	"smallint":            "int32",
-	"smallint unsigned":   "uint32",
-	"int":                 "int32",
-	"int unsigned":        "uint32",
+	"smallint":            "int64",
+	"smallint unsigned":   "int64",
+	"int":                 "int64",
+	"int unsigned":        "int64",
 	"bigint":              "int64",
-	"bigint unsigned":     "uint64",
+	"bigint unsigned":     "int64",
 	"varchar":             "string",
 	"char":                "string",
 	"date":                "datatypes.Date",
 	"datetime":            "time.Time",
 	"bit(1)":              "[]uint8",
-	"tinyint":             "int32",
-	"tinyint unsigned":    "uint32",
-	"tinyint(1)":          "int32",  // tinyint(1) 默认设置成bool
-	"tinyint(1) unsigned": "uint32", // tinyint(1) 默认设置成bool
+	"tinyint":             "int64",
+	"tinyint unsigned":    "int64",
+	"tinyint(1)":          "int64", // tinyint(1) 默认设置成bool
+	"tinyint(1) unsigned": "int64", // tinyint(1) 默认设置成bool
 	"json":                "datatypes.JSON",
 	"text":                "string",
 	"timestamp":           "time.Time",
@@ -34,8 +34,8 @@ var TypeMysqlDicMp = map[string]string{
 	"double unsigned":     "float64",
 	"mediumtext":          "string",
 	"longtext":            "string",
-	"float":               "float32",
-	"float unsigned":      "float32",
+	"float":               "float64",
+	"float unsigned":      "float64",
 	"tinytext":            "string",
 	"enum":                "string",
 	"time":                "time.Time",
@@ -51,15 +51,15 @@ var TypeMysqlMatchList = []struct {
 	Key   string
 	Value string
 }{
-	{`^(tinyint)[(]\d+[)] unsigned`, "uint32"},
-	{`^(smallint)[(]\d+[)] unsigned`, "uint32"},
-	{`^(int)[(]\d+[)] unsigned`, "uint32"},
-	{`^(bigint)[(]\d+[)] unsigned`, "uint64"},
+	{`^(tinyint)[(]\d+[)] unsigned`, "int64"},
+	{`^(smallint)[(]\d+[)] unsigned`, "int64"},
+	{`^(int)[(]\d+[)] unsigned`, "int64"},
+	{`^(bigint)[(]\d+[)] unsigned`, "int64"},
 	{`^(float)[(]\d+,\d+[)] unsigned`, "float64"},
 	{`^(double)[(]\d+,\d+[)] unsigned`, "float64"},
-	{`^(tinyint)[(]\d+[)]`, "int32"},
-	{`^(smallint)[(]\d+[)]`, "int32"},
-	{`^(int)[(]\d+[)]`, "int32"},
+	{`^(tinyint)[(]\d+[)]`, "int64"},
+	{`^(smallint)[(]\d+[)]`, "int64"},
+	{`^(int)[(]\d+[)]`, "int64"},
 	{`^(bigint)[(]\d+[)]`, "int64"},
 	{`^(char)[(]\d+[)]`, "string"},
 	{`^(enum)[(](.)+[)]`, "string"},
@@ -74,5 +74,5 @@ var TypeMysqlMatchList = []struct {
 	{`^(datetime)[(]\d+[)]`, "time.Time"},
 	{`^(bit)[(]\d+[)]`, "[]uint8"},
 	{`^(text)[(]\d+[)]`, "string"},
-	{`^(integer)[(]\d+[)]`, "int32"},
+	{`^(integer)[(]\d+[)]`, "int64"},
 }
